Don't include  in browser build
This addresses an OOM in Saucelabs, since  tries to make grpc-js work in IE11. We should not include any tests that depend on GRPC-JS
