+++ packages/firestore/test/unit/bootstrap.ts
@@ -29,5 +29,5 @@ import '../../src/platform_browser/browser_init';
 const testsContext = (require as any).context('.', true, /.test$/);
 const browserTests = testsContext
   .keys()
-  .filter((file: string) => file.indexOf('/node/') < 0);
+  .filter((file: string) => !file.match(/([\/.])node([\/.])/));
 browserTests.forEach(testsContext);
