+++ packages/firestore/test/unit/remote/serializer.browser.test.ts
@@ -1,0 +1,3 @@
+import {serializerTest} from "./serializer.helper";
+
+serializerTest();
